+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 2.2.3 (unreleased)
 ------------------
 
+Features and improvements:
+
 - ``wait_until`` and ``wait_until_not`` now accept commands with no ``sub_commands`` property
 
 2.2.2 (2019-03-29)